Added $repo_class and $web_cmd parameters.
About $repo_class. Requiring `yum` class should be optional and certain
companies already have (custom) classes that do that. In such environments
it's useful to allow the user to choose which class to use. This parameter
can be `undef` (in this case no class is included).

About $web_cmd. Some companies do not allow servers to access directly
the Internet but they require a proxy. Being able to define the $WEB_CMD
config variable helps in this kind of situation.